Add some basic tests for the ROBOT plugin.
Add a basic test for:

* sssom:inject's --direct option (direct transformation of a mapping set
  to OWL);
* sssom:xref-extract;
* sssom:rename.

Those tests are far from a "good" coverage of their corresponding
features, but they will do for. Those features are quite "frozen" anyway
(for now at least), so they are unlikely to be inadvertently broken.
We'll improve the coverage if/when we are planning to bring some changes
to those features, as we have done for SSSOM/T.