Move all non-SSSOM-core features into sssom-ext.
Create a new module named 'sssom-ext' and move all the code related to
SSSOM/Transform, OWL features, and support for URI Expressions into the
new module.

The idea is that 'sssom-core' is strictly about implementing the SSSOM
specification (i.e. providing tools to read, write, and perform basic
manipulations on mappings and mapping sets). SSSOM/Transform and
OWL-related features, and (experimental) support for URI Expressions
should be distinct from that.