Build the site at the same as the package.
Now that the website is split over several inter-dependent modules, the
packages must be built as part of the same Maven invocation as the
website (trying to run 'mvn site' alone would fail because Maven would
not be able to find the module that another module is dependent on).

So we can and in fact must build everything in the same step.