Allow using is_a on object and data properties.
Extend the role of the SSSOM/T-OWL function "is_a" so that it can be
used to check whether an object or data property is a subproperty of
another property (instead of only being usable to test classes).

Also make it possible to use the same function to test whether a given
entity is a class, an object property, a data property, or an annotation
property.