Annotate slots with their slot URIs.
If a SSSOM slot is associated, in the SSSOM LinkML model, with a URI
that is distinct from the default URI (sssom:slot_name), we add an
annotation to the corresponding field so that the URI is accessible to
Java code at runtime.

The Slot class gets a new method `getURI()` to allow easy access to a
slot URI (whether it is a default URI or an explicitly specified URI).

Relatedly, we also annotate fields corresponding to slots that have a
'uri' range with a dedicated annotation.