Support serialisation of non-standard metadata slots in RDF.
Implement serialisation of extension slots and extension definitions in
the RDF output format.

Of note, the IRIs to use for representing the fields of the 'extension
definition' structure are _not_ defined in the spec. The output of the
serialiser for a mapping set containing defined extensions should be
considered experimental and is subject to change.