Do not write extension definitions in RDF by default.
When saving the result set in RDF, by default, configure the writer to
write extensions in UNDEFINED mode -- that is, write extension slots and
their values but do not write the extension definitions.

Extension definitions are not really useful in RDF. In SSSOM (TSV or
JSON), their main purpose is to provide the property IRI that gives a
meaning to an extension slot. But in RDF, the property IRI is already
present no matter what (the extensions are serialised using their
associated IRIs), so there should be no real need for extension
definitions.

People who want RDF output with extension definitions can still use the
`--write-extra-metadata=DEFINED` option to force the writer to output
them.