Strip trailing tabs in embedded YAML metadata block.
When a SSSOM/TSV file with an embedded YAML metadata block is modified
with a common spreadsheet editor (e.g. LibreOffice Calc, Microsoft
Excel, or Apple Numbers), upon saving the editor may append trailing
tabs at the end of every line in the header (that's the behaviour
observed with all the aforementioned editors). Those trailings tabs need
to removed, otherwise they will cause the YAML parser to fail to read
the metadata block.