Add util package to facilitate reading from all supported formats.
Add a new org.incenp.obofoundry.sssom.util package providing a
ReaderFactory class that can be used to try to automatically get a SSSOM
reader object for any supported serialisation format. The factory will,
if possible, look at the first byte of a file to figure out the format
used (SSSOM/TSV, SSSOM/JSON, or RDF/Turtle) and return a suitable
reader.

Update SSSOM-CLI and the ROBOT plugins to use the new factory, so that
they both gain the ability to read RDF/Turtle files in addition to
SSSOM/TSV and SSSOM/JSON.