Rename BaseReader/BaseWriter to SSSOMReader/SSSOMWriter.
This is, strictly speaking, a breaking change, but one that should not
have a big impact. Until the introduction of the ReaderFactory in the
Extended Library, there was little reason for client code to refer to
the BaseReader class directly, and there is still little reason for
referring to the BaseWriter class. Client code would rather use
TSVReader/JSONReader or TSVWriter/JSONWriter directly.

Now that there _is_ a reason to explicitly refer to the base class, at
least for the reader, when using the ReaderFactory:

    BaseReader reader = readerFactory.getReader(myFile);

and so the name BaseReader, which was fine as long as that class was
merely an internal implementation detail, is not so fine anymore.