Re-design the slot and slot visitor model.
This commit removes the SlotVisitor interface and replaces it with a new
ISlotVisitor that works slightly differently. Mainly, the "visit"
methods are now distributed based on the _SSSOM_ slot type, rather than
the _Java_ field type.

This means that instead of having, for example, one method to visit all
slots that are backed by a Java String, there is now three distinct
methods: one to visit actual String-typed (in SSSOM sense) slots, one to
visit URI-typed slots, and one to visit EntityReference-typed slots --
even though all of these slots are represented in Java by a String
field.

To support this, the Slot class is now derived into as many subclasses
as there are SSSOM slot types, so we now have a StringSlot class, a
URISlot class, a EntityReferenceClass, a DateSlot class, etc.

In addition, the new interface defines methods with a fixed return type
of "void", instead of having a parameterized return type. This is
because a parameterized return type was found not to be really useful in
practice.

The removal of the old SlotVisitor interface is clearly a breaking API
change, but since we are only a couple of weeks post-1.0 this is one I
consider acceptable, since the new model is, I believe, much cleaner and
I don't want to keep the old one around, even marked as deprecated.