Move all slot-manipulation code to the new slots package.
Since we already introduced a breaking API change with the new slot
visitor, let's go all the way and cleanup the top-level package a little
bit. There are two many classes in that package, so we move all the code
intended to facilitate working with metadata slots to the new
org.incenp.obofoundry.sssom.slots package. The top-level package is now
mainly intended for I/O code.