Add support for `predicate_type`.
Recognise the `predicate_type` slot in mappings and mapping sets. Check
that it does not contain the illegal values `rdfs literal` and `composed
entity expression`.

Add a `predicate_type==` filter in SSSOM/T.

When producing OWL axioms from mappings in "direct" mode, use the value
of the `predicate_type` slot, if present, to determine the type of axiom
to generate.