Allow using the ExtensionSlotManager without a PrefixManager.
When we build a SSSOM mapping set from a RDF model, we always have
already expanded, full-length identifiers at our disposal (even if the
RDF model was stored in a format that makes use of shortened
identifiers), so we don't need the ExtensionSlotManager to have a
PrefixManager to take care of the identifier expansion.

So we make the use of a PrefixManager with the ExtensionSlotManager
optional: add a new constructor without a PrefixManager argument, and
guard all uses of the PrefixManager with nullness checks.

Also add a method to get an extension definition from a property name
(which is all that we will have, when we'll be processing extension
slots from a RDF model) rather than from a slot name.