Allow several XSD URIs for a given extension ValueType.
There are several XSD types that are suitable to represent an integer
value: beyond the generic xsd:integer, there is also xsd:int, xsd:long,
xsd:short, xsd:byte...

Likewise, a string type can be represented e.g. by xsd:normalizedString
in addition to xsd:string.

So, we generalise the concept of "alias" (which we originally used to
allow both linkml:Uriorcurie and linkml:uriOrCurie to mark an
identifier) and allow any ValueType value to be represented by an
arbitrary number of IRIs: a "canonical" IRI and some aliases.