Allow condensation/propagation in RDF.
Allow slot propagation when reading from RDF, and slot condensation when
writing to RDF.

I don't think slots should ever be condensed when writing to RDF, as
propagation is a SSSOM-specific concept and therefore a RDF tool would
likely not know what to do with condensed slots. But it's easier to
treat all serialisation formats in the same way. We can disable
condensation by default at the level of SSSOM-CLI.