Disable condensation by default when writing to RDF.
When the output format is set to RDF/Turtle, we disable condensation of
slots by default. RDF-consuming tools would likely be unaware of the
very concept of propagatable/condensable slots (which is highly specific
to SSSOM), so writing the set in "uncondensed" form is the most sensible
default in that format.

Explicitly setting the --condensation (or --condensation=true) option
will force condensation to be enabled even in RDF output.