Try inferring the output format by default.
If no output format is explicitly specified with the --output-format
option, try using the extension of the output file to automatically
determine the output format. If the extension is not recognised, or if
no output file was specified (writing to standard output), keep using
SSSOM/TSV as the default output format.