Accept SSSOM/TSV files with no metadata block.
In principle, the metadata block in the SSSOM/TSV format is intended to
be *mandatory* (independently of whether it is embedded or external),
but the spec also says that empty metadata slots SHOULD be omitted. It
is therefore strictly speaking allowed for the metadata block to be
completely absent, for a mapping set whose ALL metadata slots are empty.

So we update the TSV reader to silently accept a SSSOM/TSV file with no
metadata block and no accompanying external metadata file -- provided of
course that the mapping set does not require a CURIE map.