Add read support for "SSSOM/CSV".
Make it possible to read a "SSSOM/CSV" file, assuming the format is
exactly identical to SSSOM/TSV *except* for the use of commas rather
than tabs to separate column.

The SSSOM/CSV format is not officially part of the SSSOM spec, but
SSSOM-Py silently supports it (`sssom validate` does not even emit a
warning ><) and as a consequence people have started using it in the
wild, so we better support it as well.