Add write support for the SSSOM/CSV format.
Add the possibility to write SSSOM/CSV file. From the client code's
point of view, this works in a similar fashion as reading: the client
should create a TSVWriter instance and call the enableCSV() method on
it.

Since SSSOM/CSV is now supported both for reading and writing, we add it
to the SerialisationFormat enum in the extended library, which allows us
to (1) simplify the inferFormat code in the ReaderFactory and (2) make
that format available to the --output-format option of SSSOM-CLI.