Try to guess whether a SSSOM/TSV file is really TSV or CSV.
Before reading the TSV section of a SSSOM/TSV file, peek into the first
32 bytes of the header line to try and guess the column separator (tab
or comma). If we do not find any separator (which should be unlikely if
the file is truly a SSSOM file, as no SSSOM slot has a name longer than
32 characters), we default to expecting tabs.

The reader can be forced to expect either tabs or commas (instead of
attempting to guess the separator) by calling the setSeparatorMode
method.

When obtaining a reader from the ReaderFactory of the extended library,
if the filename's extension (if available) indicates that a file is a
SSSOM/TSV (resp SSSOM/CSV) file, then the reader will be configured to
expected tab (resp comma) separators. Even if we accept indifferently
SSSOM/TSV and SSSOM/CSV files, it is an error to have comma-separated
columns in a file carrying a '.sssom.tsv' extension (or inversely to
have tab-separated columns in a file with a '.sssom.csv' extension).