Allow not to sort the mappings when writing.
Mappings are written in a deterministic order, as recommended by the
SSSOM specification.

Here, we do the sorting in the base SSSOMWriter class (before calling
the abstract `doWrite()` method), so that mappings are sorted regardless
of the actual output format (this makes the JSONWriter automatically
sorts the mappings, which it previously did not do).

We also add the possibility to skip the sorting step, should client want
to do that.