Allow using standard-specified properties when annotating axioms.
When producing OWL axioms, they can be annotated using the metadata from
the mappings. By default, a given metadata field, when present, is
translated to an axiom annotation using an annotation property
fabricated using the SSSOM IRI of the metadata field.

For example, a `creator_id` metadata field would yield an annotation
using a property with the IRI `https://w3id.org/sssom/creator_id`.

This is the normal "OWL reification" rule, as mandated by the SSSOM
specification.

But some SSSOM metadata slots are associated with another IRI (specified
in the schema using the `slot_uri` LinkML slot), and sometimes it may be
desirable to use those other IRIs rather than the SSSOM IRIs.

For example, `creator_id` is associated with the
`http://purl.org/dc/terms/creator` IRI, so it may be desirable to
annotate an axiom using that property IRI rather than
`https://w3id.org/sssom/creator_id`.

For that, we introduce a new IMetadataTransformer implementation that
allows to obtain the suitable IRI for any given slot. And in the
SSSOM/T-OWL language, we introduce a new named parameter in all
axiom-producing functions, `/annots_uris=`, which indicate how
annotation IRIs should be obtain. That parameter can take two values:
`direct`, which is the default behaviour explained above (where a
`creator_id` field would yield a `ssssom:creator_id` annotation), and
`standard_map`, which is the behaviour using the mapped IRIs defined in
the SSSOM schema (where `creator_id` would yield a `dcterms:creator`
annotation; for a slot that does not have a mapped IRI, the SSSOM IRI is
used instead, as in the "direct" mode). The default behaviour is the
"direct" one, for backwards compatibility.