Fix EPM-based IRI canonicalisation.
The canonicalisation code did not properly handle the following case:

Consider the IRI <http://purl.obolibrary.org/obo/UniProtKB_1234>. The
<http://purl.obolibrary.org/obo/UniProtKB_> is, in the OBO extended
prefix map, a prefix synonym for the canonical prefix
<http://purl.uniprot.org/isoforms/>:

```
{
    "uri_prefix": "http://purl.uniprot.org/isoforms/",
    "uri_prefix_synonyms": [
            "http://purl.obolibrary.org/obo/UniProtKB_"
    ]
}
```

but because we are looking for a canonical prefix before looking up for
a synonym, we find that <http://purl.obolibrary.org/obo/UniProtKB_1234>
is already a canonical IRI within the <http://purl.obolibrary.org/obo/>
namespace.

We should not lookup for "canonical" and "synonym" IRI prefixes
separately, but in a single step instead, so that we can always find the
longest matching prefix, regardless of whether it is a canonical one or
a synonym.