Add the --epm-mode option to SSSOM-CLI.
Add an option to SSSOM-CLI to specify how and when the extended prefix
map (if any) is to be used.

There are three different modes.

In PRE mode, the EPM is used _before_ the input set is read, to
complement the input set's own prefix map (allowing to read a set even
if it has an incomplete map).

In POST mode, the EPM is used _afer_ the input set is read to reconcile
the set's IRIs so that they only use canonical IRI prefixes.

The third mode, BOTH, is a combination of the PRE and POST modes. This
is the default mode.

The old --mangle-iris option is renamed to simply --epm. The old name is
supported for backwards compatibility. `--mangle-iris FILE` is
equivalent to `--epm FILE --epm-mode POST`.