Add "initialising read accessors".
For fields represented by "complex" types (list or maps), that need to
be initialised before they can be used, we add a `get` accessor that can
optionally initialise the field before returning its value.

For example, to add an "author label" to a mapping, it is now possible
to do

  mapping.getAuthorLabel(true).add("A. U. Thor");

which will always work even if the `authorLabel` has never been set
(either by calling `setAuthorLabel()` or using the `authorLabel()`
method of the MappingBuilder object).