Add Javadocs comments to custom generated methods.
We can't add Javadocs comments to LinkML+Lombok-generated methods, but
we can at least add comments to the methods we are custom-generating.
This is especially important for the `getX(boolean set)` accessors,
since it is not obvious what the argument is for.

Also avoid generating a Java class for the `NoTermFound` thing, which is
defined as a class in LinkML (no idea why, but that's LinkML for you)
but certainly does not need to be.