Add a method to get the minimum SSSOM version required by a set.
Add to the Validator class (which may not be the best place for that,
but that will do for now) a method to inspect an entire set and
determine the minimum version of the spec the set is compliant with.

This will be needed by the writer objects to automatically fill the
sssom_version slot.