Support reading and writing the sssom_version slot.
Update the TSV and JSON readers and writers to make them support the
sssom_version slot.

When reading, the sssom_version slot is checked first, and its value
will determine whether post-1.0 slots are recognised or not.

When writing, the value of the sssom_version slot in the set to write is
ignored, and replaced by a value computed on the fly based on the slots
that are effectively used in the set.