Do not process the sssom_version slot twice.
Since we look for the sssom_version slot right at the beginning of the
YAML conversion process, we can (and must) remove it from the rawMap, so
that it will not be treated a second time by the loop that processes all
normal slots.

The second processing would not be an issue for a SSSOM 1.1+ set (it
would be useless but harmless), but for a SSSOM 1.0 set, it could cause
the sssom_version to be recognised as an extension slot (which, strictly
speaking, would be correct, but no really useful).