Allow assuming a different SSSOM version.
The SSSOM specification mandates that, if the sssom_version slot is not
present, a set must be assumed to be compliant with version 1.0.

However it might be useful to optionally change that assumption, and
assume that a set is compliant with another version. For that, we add a
`#setAssumedVersion()` method to the base `SSSOMReader` class, to set
the version a set should be assumed to be compliant with in the absence
of an explicit `sssom_version` slot.

At the CLI level, we add a `--assume-version=` option to `sssom-cli`.