Allow ignoring missing imports when updating from an ontology.
Add a `--ignore-missing-imports` option to allow the OWLAPI to silently
ignore any missing imports when loading an ontology.

This is mostly a workaround for what seems to be a (very annoying) bug
in Java 11 (also observed with Java 17, but not with Java 21): if the
XML catalog contains a `group`, then all entries _after that group_ are
ignored. That is:

  <catalog ...>
    <group ...>
      <!-- All entries here are correctly taken into account -->
    </group>
    <!-- All entries here are ignored -->
  </catalog>