Check for correctness of URI-typed slots.
URI-typed slots, even if they are represented as plain strings, should
contain valid URIs, so we get both the YAMLConverter (used by the TSV
and JSON parsers) and the RDFConverter (used by the RDF parser) to check
the value of URI-typed slots, and throw a SSSOMFormatException if they
do not contain a valid URI.

In addition, in SSSOM 1.1-compliant sets, it will be expected that the
URI must be absolute and not relative -- in SSSOM 1.0, relative URIs
must still be accepted, because (1) formally the SSSOM 1.0 spec did not
make clear that absolute URIs were expected, and (2) SSSOM-Py has always
accepted relative URIs without complaining.

Checking for correctness of URI values requires updating the
'test-escaping-yaml.sssom.tsv' test case to use a non-URI-typed slot
when playing with various YAML corner cases.