Use mapping_set_id as the RDF resource representing the mapping set.
When serialising to RDF, if the mapping set has a `mapping_set_id`
(which it should, in principle, always have, as it a _required_ slot),
then we use that ID as the resource representing the set, instead of a
blank node.

That is, instead of serialising the set as

  [] a sssom:MappingSet ;
     sssom:mapping_set_id "https://example.org/set"^^xsd:anyURI .

we serialise as

  <https://example.org/set> a sssom:MappingSet .

But we still support the blank node form when deserialising from RDF,
for backward compatibility.