sssom-cli: Add `--lax` option.
We add a `--lax` option to SSSOM-CLI to allow the tool to silently
accept sets that it would otherwise reject because they are, strictly
speaking, invalid.

Internally, this is done by introducing the concept of "validation
level", where different kinds of checks are performed at each level. The
default behaviour is to perform a "full" validation, where the slightest
deviation from the spec requirements causes a set to be rejected as
invalid. The `--lax` option triggers instead a "minimal" validation,
where we only check that the absolutely required `subject_id`,
`predicate_it`, `object_id`, and `mapping_justification` slots are
always present for every mapping.