Add method to create a canonical S-expression of a mapping.
Add an automatically generated Mapping#toSExpr() method to create a
canonical S-expression representing a mapping. The S-expression takes
into account all non-null slots, except the `record_id` slot.

It is primarily intended to allow comparing mapping instances to
determine whether they are identical (siimilar to #toString(), but
ignoring the value of `record_id`, if any) and to compute a hash from a
mapping.