Add the `default` substitution modifier function.
Add a new substitution modifier function called `default`, that can be
used to insert a default value if the original substituted value happens
to be null or empty.

This changes the default behaviour of the MappingFormatter to return an
empty string, rather than the original formatting string, if the
formatting string does not correspond to any known slot or extension.