Ignore empty-valued extensions.
When deserialising a set (especially from TSV), if an extension slot has
an empty value (for example if the corresponding column in the TSV file
is empty), ignore the extension slot entirely. That is, do not add an
ExtensionValue object which does not contain an actual value.

This allows to check whether a mapping contains a particular extension
just by checking for the property in the extensions map:

  if ( mapping.getExtensions().containsKey("my_extension") ) { ... }

rather than having to also check whether the ExtensionValue contains an
actual value:

  ExtensionValue ev = mapping.getExtensions().get("my_extension");
  if ( ev != null && ev.getValue() != null ) { ... }