Fix validation of record IDs.
The way we were checking that either all records within a set have a
record ID or none of them have such a ID was flawed and did not handle
all cases correctly. In particular, it missed the case where only the
*last* record of a set had an ID.

We now proceed by checking whether the *first* record has an ID or not,
and comparing all other records to the first: if the first record has an
ID, then they should all have one; if it doesn't have an ID, then none
of the other records should have one.