Add the SetMerger class.
Replace the MetadataMerger class in SSSOM-CLI by a more generic
SetMerger class in SSSOM-Core.

The new class can be used to merge both mappings and metadata and can be
more finely controlled using a set of MergeOption.

Of note, the test case for the new SetMerger class is simply a
transposition of the the case for the old MetadataMerger class. As such,
it only tests the _default_ behaviour of the SetMerger. More tests will
be needed to cover the alternative behaviours that are now possible.