Recognise a TTL file starting with a non-blank resource.
When peeking into a file to infer its format, we currently recognise a
TTL file by checking whether it starts with `@` (a `@prefix...`
declaration) or with `[` (`[] a <https://w3id.org/sssom/MappingSet>`).

But since we now expect that the `mapping_set_id` will be used as the
resource IRI, if the TTL file does not use CURIEfied identifiers and
therefore does not start with `@prefix` declarations then it will start
with something like `<my.set.id> a <http://www.w3.org/2002/07/owl#>`.

So we recognise a file starting with `<` as a TTL file.