Implements the proposed "extraction syntax" feature.
Add a feature (proposed to be added to the core spec) allowing to
extract an arbitrary piece of content from a mapping set, using a string
to "locate" said piece of content relatively to a mapping set object.

Usage is as follows:

    ValueExtractorFactory fac = new ValueExtractorFactory();
    IValueExtractor extractor = fac.parse("mapping(2).slot.subject_id");
    Object value = extractor.extract(mappingSet);

where

    mapping(2).slot.subject_id

instructs to extract the `subject_id` slot from the second mapping of
the set.

Documentation for the full syntax of locator expressions will come
later, once the proposal has been discussed and hopefully accepted in
the SSSOM spec.