Factorise the code to format double- and date-typed values.
Add a new helper class providing static methods to perform some common
tasks. For now, this class provides methods to format double- and
date-typed values according to the recommendations of the SSSOM
specifications.

Update the TSVWriter and JSONWriter classes to use the new helper
methods. Same for the #toString() method of the ExtensionValue class.