sssom-cli: Add the --extract option.
Add a new option to sssom-cli to extract a single value from anywhere
within the result set, using the newly proposed "extractor expression"
syntax.

The option takes a single argument which is the expression indicating
the value to extract, e.g:

   sssom-cli <INPUT> --extract mapping(2).slot.subject_id

to extract the `subject_id` of the second mapping.

Another option allows to disable writing the result set to the standard
output. This is independent from the `--extract` feature, but would
typically be used at the same time, so that the command only prints the
extracted value and _not_ the full result set.