Tweak and document the extractor syntax.
Change slightly the syntax of extractor expressions, to make the second
component always look like a "function call".

That is, instead of

  set.slot.subject_id

we now use

  set.slot(subject_id)

Likewise for `mapping.special.hash` and `mapping.special.sexpr`, which
become `mapping.special(hash)` and `mapping.special(sexpr)`.

This avoids having two different syntaxes, one for extensions and one
for slots and special values.