rdf: Write URI-typed slots as named resources.
The RDF serialisation of SSSOM is still being discussed, but one aspect
for which there is seemingly agreement is that URI-typed slots should be
rendered as IRIs, rather than as xsd:anyURI literals.

So we change the behaviour of the RDF writer to do precisely that.

The RDF reader still accepts both xsd:anyURI literals and IRIs for
URI-typed slots.