Allow full export to OWL.
Add helper methods to (1) annotate an ontology with annotations derived
from the metadata of a mapping set, and (2) create an new ontology
entirely derived from a mapping set, representing an "OWL export" of the
mapping set (where ontology annotations represent the set's metadata,
axioms represent the mappings, and axiom annotations represent the
mappings' metadata).

In the ROBOT plugin's `sssom:inject` command, the combined use of the
`--create` option (to create a new ontology) and of the `--direct`
option (to create a direct OWL translation of the mappings) is
interpreted that the newly created ontology should be annotated with
set-derived annotations.