Change default SSSOM/T-OWL metadata transformer.
Change the behaviour of SSSOM/T-OWL functions so that, when annotating
axioms from SSSOM metadata, the annotation properties are obtained from
the "uri" fields of the SSSOM model.

This means that the annotation properties used in OWL reification are
the same as those used in RDF, which is (probably) what was intended
even though this is not what is shown in the examples currently given by
the spec.

For example, a `creator_id` slot is derived into a `dcterms:creator`
annotation, instead of a `sssom:creator_id` annotation.

Old-style direct annotations (e.g. `sssom:creator_id` instead of
`dcterms:creator`) can still be obtained by using the flag parameter
`/annots_uris="direct"` (which was the old default).