sssom-cli: Add the --input-format option.
Automatically inferring the format from the filename or from its content
is nice, but there may be cases when the format cannot be automatically
detected, and so being able to explicitly specify it may be useful.

For example, the detection logic currently fails to correctly detect
Turtle files that use Turtle 1.1 case-insensitive `PREFIX` declarations
-- it can only detect Turtle if the prefix declarations start with a
`@` character, as in Turtle 1.0.

This does not cause any issue if reading from a file as long as the file
as a `.ttl` extension, but this will cause an issue if reading from
standard input. For example, the following will fail:

  sssom-cli myset.sssom.tsv --output-format ttl | \
    riot --syntax turtle --output turtle | \
    sssom-cli

because the second `sssom-cli` cannot recognise the output of `riot`
(which follows Turtle 1.1 syntax) as a Turtle file, and therefore
defaults to assuming it is a TSV file.

The following will roundtrip as expected:

  sssom-cli myset.sssom.tsv --output-format ttl | \
    riot --syntax turtle --output turtle | \
    sssom-cli --input-format ttl