Support rendering URI-typed slots as IRIs in OWL serialisation.
In SSSOM/RDF, URI-typed slots are rendered as named resources rather
than xsd:anyURI literals. Presumably, when exporting to SSSOM/OWL the
behaviour should be similar. However, this has not been decided yet (the
SSSOM/OWL serialisation is still largely unspecified), so for now we
support rendering URI-typed slots as both xsd:anyURI literals and IRIs,
with an internal option to switch from one behaviour to the other.