Add new Cardinalizer class.
There is currently no precise and agreed upon definition of
"cardinality" -- the definition in the current spec is precise, but not
agreed upon as the original author himself is unsure about what the
definition should be.

We are heading towards a more flexible definition of cardinality, where
cardinality is always relative to a "scope" -- a list of all the slots
that must be taken into account when computing cardinality values.

For example, cardinality relative to the (predicate_id, object_source)
slot is computed on all mappings that have the same predicate and the
same object source.

To cope with the new flexible definition, we deprecate the simple
`inferCardinality` static method in the `MappingCardinality` enum, and
replace it with a new class entirely dedicated to cardinality
computation.