Disable cardinality tracking when infer_cardinality is used.
The SSSOM/T processor normally takes care of (1) computing the effective
cardinality of mappings whenever a `cardinality==` filter is used, and
(2) tracking the changes to the input set in order to know whether
previously computed cardinality values are still valid or not.

When it does compute cardinality, the processor always compute the
cardinality relative to the entire set, which may not be what users
need. Therefore, users can now use the `infer_cardinality` function to
control themselves (1) when cardinality is computed and (2) which
cardinality (relative to what) is computed. But for that to work, we
must inhibit the default behaviour of automatically tracking and
computing the cardinality.

So to both allow users to control cardinality computation if they so
need, and preserve backwards compatibility, the exact behaviour of the
SSSOM/T processor depends on whether the `infer_cardinality` function is
used or not:

* if it is not used, the current behaviour of automatically computing
  the cardinality relative to the entire set is preserved;
* conversely, if `infer_cardinality` is used even once anywhere in the
  ruleset, then we assume that the ruleset takes responsibility for
  computing cardinality, and the default behaviour is inhibited.