Support cardinality_scope.
Implement support for the new `cardinality_scope` slot.

Amend the behaviour of the Cardinalizer class to automatically fill that
slot at the same time as the `mapping_cardinality` slot -- this also
automatically updates the behaviour of the SSSOM/T `infer_cardinality`
function, since it relies on the Cardinalizer.

Add a `cardinality_scope` filter in SSSOM/T.