Correctly recognise slots newly added to the MappingSet class.
It may happen that a SSSOM slot has existed since version 1.0 of the
specification (or even before), but has only been added to the
MappingSet class in a more recent version.

For such slots, we need to know when they have been added to the
MappingSet class, in order to correctly determine the minimum version a
given set is compliant with.

We do that by forcefully adding a `addedIn` annotation to such slots in
the MappingSet class. This is used additional knowledge hardcoded in the
custom Java generator, since the LinkML model cannot provide us with the
required information.