Extract all addedIn annotations from the LinkML model.
There is a way (not used yet in the SSSOM LinkML schema) to encode in
the schema the fact that a given slot has been added to a class in a
specific version of the model.

Assuming the SSSOM schema will adopt that method, we can now extract
that information from the schema, instead of having to hardcode it in
the generator script.