Accept non-ambiguous invalid enum values.
When reading from SSSOM/TSV or SSSOM/JSON, accept a `sssom_version`
value provided as a floating point value, even though strictly speaking
it should be a string. That is, accept `sssom_version: 1.1` as
equivalent to `sssom_version: "1.1"`.

Likewise, accept `entity_type_enum` values provided as CURIEs. For
example, accept `owl:Class` as equivalent to `owl class`.

In both cases, this is when parsing only. When writing, the version
number is always written as a string and entity type values are alwayrs
written using their canonical string form.