Allow condensing for a given SSSOM target version.
Add a `setMaxVersion()` property to the SlotPropagator class, to define
the highest version of the SSSOM version of the spec that sets should
remain compliant with even after condensation.

For example, the `curation_rule` slot did not exist on the MappingSet
class in version 1.0; it was only added (and made propagatable) in
version 1.1. Therefore, a 1.0-compliant set where all mappings have the
same `curation_rule` would by default be condensed into a set where the
`curation_rule` value is at the level of the set, and such a set would
no longer be compliant with 1.0.

Setting a maxVersion of 1.0 will instruct the SlotPropagator class _not_
to condense the `curation_rule` slot, so that the condensed set remains
compliant with version 1.0.