Add the raw_predicate filter.
Because the `predicate==...` filter always take the predicate modifier
into account, there is currently no way of selecting mappings that have
a given predicate with the `Not` modifier.

For example, if one wants to select the mappings that have a negated
`skos:exactMatch` predicate:

* `!predicate==skos:exactMatch` will not work, because the inner filter
  will never select the mapping because of the Not modifier;

* `!predicate==skos:exactMatch && predicate_modifier==Not` will select
  all negated mappings, even those with a different predicate.

So we introduce a new predicate filter, called `raw_predicate`, that
purposefully _ignores_ the predicate modifier, and instead only checks
the value of the predicate_id slot. This makes it possible to do

  `raw_predicate==skos:exactMatch` && predicate_modifier==Not

to precisely select the negated skos:exactMatch mappings.