Add option to split output according to predicates.
This commit adds a `--split-with-predicates` flag which changes the
behavior of the `--split` flag. Typically, the output is split along a
subject prefix and an object prefix, e.g. `MONDO-to-NCIT.sssom.tsv`.
When the `--split-with-predicates` flag is passed, it also includes the
CURIE of the relation in the name of the output split file, e.g.
`MONDO-skos_exactMatch-NCIT.sssom.tsv`. (Note that the colon in the
CURIE has been replaced by an underscore, since colons cannot appear in
filenames in Windows).

This (almost) matches the behavior of sssom-py:

<https://github.com/mapping-commons/sssom-py/blob/ac0f769386d35e2b1c961cac59221c816143bf07/src/sssom/parsers.py#L1033-L1035>

The difference being that sssom-py only includes the local name of the
predicate, while this pull request includes the full CURIE.