Add `--split-by` option.
Add an option to split a mapping set along an arbitrarily specified
format string. The format string reuses the same syntax as used for
argument substitutions in the context of SSSOM/T.

The default behaviour of the `--split` option, and of the newly added
`--split-with-predicates` option, are equivalent to the following two
format strings, respectively:

  %{subject_id|prefix}-to-%{object_id|prefix}

and

  %{subject_id|prefix}-%{predicate_id|short|replace(':', '_')}-%{object_id|short}

The format modifiers `prefix` (to get only the prefix part of a
shortened identifier) and `replace` (to perform arbitrary
find-and-replace operations within a string) are added in this commit.