Centralize registration of standard modifier functions.
Add a `setStandardModifiers()` method to the `MappingFormatter` class to
automatically register the modifier functions that are provided within
SSSOM-Java. Now that the mapping formatter is used in two different
places (within the SSSOM/T engine and within SSSOM-CLI's `--split-by`
feature), this allows to ensure that the same modifier functions are
always available in both places.