Refactor format modifier functions.
Most format modifier functions behave the same way when they are called
on a list of values rather than a single value: they simply process each
value in the list in the same way as they would process a single value.

The code to deal with list of values was duplicated in every single
modifier functions, so here we insert a base class with that common
logic implemented once and for all.